requirements.txt
# Core Framework - Updated for Python 3.13+ compatibility
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
python-multipart==0.0.20

# Twilio
twilio>=9.0.0

# AI Services
google-genai>=0.3.0

# Google Maps
googlemaps>=4.10.0

# Configuration - Pydantic v2
python-dotenv>=1.0.0
pydantic==2.10.6
pydantic-settings==2.7.1

# HTTP Client
aiohttp>=3.11.0